refactor: split portable decode stages out of file.rs (source-quality ceiling)
P4's estimator/ceiling wiring pushed file.rs to 2547 lines against the 2500
ceiling; topology decode, optional-section load, and the norm rebuild move
to io/file/portable_sections.rs unchanged.
